+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>gradle</name>
+	<name>mdwriter</name>
 	<comment>Project gradle created by Buildship.</comment>
 	<projects>
 	</projects>
