--- app/bin/default/META-INF/extensions.idx
@@ -1,2 +1,0 @@
-# Generated by PF4J
-org.pf4j.demo.WhazzupGreeting
