--- app/src/main/java/org/pf4j/demo/WhazzupGreeting.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2012-present the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.pf4j.demo;
-
-import org.pf4j.Extension;
-import org.pf4j.demo.api.Greeting;
-
-/**
- * @author Decebal Suiu
- */
-@Extension
-public class WhazzupGreeting implements Greeting {
-
-    @Override
-    public String getGreeting() {
-        return "Whazzup";
-    }
-
-}
