+++ mdwriter-api/.settings/org.eclipse.buildship.core.prefs
@@ -2,7 +2,7 @@ arguments=--init-script /home/hancy/.local/share/nvim/mason/packages/jdtls/confi
 auto.sync=false
 build.scans.enabled=false
 connection.gradle.distribution=GRADLE_DISTRIBUTION(LOCAL_INSTALLATION(/home/hancy/.sdkman/candidates/gradle/current))
-connection.project.dir=../../app
+connection.project.dir=../app
 eclipse.preferences.version=1
 gradle.user.home=
 java.home=/usr/lib/jvm/java-21-openjdk-amd64
