+++ mdwriter-app/.settings/org.eclipse.jdt.apt.core.prefs
