+++ mdwriter-app/.settings/org.eclipse.jdt.core.prefs
