+++ mdwriter-app/bin/main/log4j.properties
