+++ mdwriter-app/src/main/resources/log4j.properties
